More work in isolating Qt functionality
