+++ glue/plugins/dendro_viewer/__init__.py
@@ -1,5 +1,2 @@
 def setup():
-    from glue.config import qt_client
-    from .qt.data_viewer import DendrogramViewer
     from .data_factory import load_dendro  # noqa
-    qt_client.add(DendrogramViewer)
