--- glue/viewers/histogram/__init__.py
@@ -1,4 +1,0 @@
-def setup():
-    from glue.config import qt_client
-    from .qt.data_viewer import HistogramViewer
-    qt_client.add(HistogramViewer)
