+++ glue/viewers/histogram/qt/__init__.py
@@ -1,1 +1,6 @@
 from .data_viewer import HistogramViewer  # noqa
+
+
+def setup():
+    from glue.config import qt_client
+    qt_client.add(HistogramViewer)
