+++ glue/viewers/scatter/qt/__init__.py
@@ -1,1 +1,5 @@
 from .data_viewer import ScatterViewer  # noqa
+
+def setup():
+    from glue.config import qt_client
+    qt_client.add(ScatterViewer)
