--- glue/viewers/table/__init__.py
@@ -1,4 +1,0 @@
-def setup():
-    from glue.config import qt_client
-    from .qt import TableViewer
-    qt_client.add(TableViewer)
