+++ setup.py
@@ -51,16 +51,17 @@ cmdclass = {}
 entry_points = """
 [glue.plugins]
 export_d3po = glue.plugins.export_d3po:setup
-export_plotly = glue.plugins.exporters.plotly:setup
+export_plotly = glue.plugins.exporters.plotly.qt:setup
 pv_slicer = glue.plugins.tools.pv_slicer:setup
 coordinate_helpers = glue.plugins.coordinate_helpers:setup
 spectral_cube = glue.plugins.data_factories.spectral_cube:setup
-dendro_viewer = glue.plugins.dendro_viewer:setup
-image_viewer = glue.viewers.image:setup
-scatter_viewer = glue.viewers.scatter:setup
-histogram_viewer = glue.viewers.histogram:setup
-profile_viewer = glue.viewers.profile:setup
-table_viewer = glue.viewers.table:setup
+dendro_factory = glue.plugins.dendro_viewer:setup
+dendro_viewer = glue.plugins.dendro_viewer.qt:setup
+image_viewer = glue.viewers.image.qt:setup
+scatter_viewer = glue.viewers.scatter.qt:setup
+histogram_viewer = glue.viewers.histogram.qt:setup
+profile_viewer = glue.viewers.profile.qt:setup
+table_viewer = glue.viewers.table.qt:setup
 data_exporters = glue.core.data_exporters:setup
 fits_format = glue.io.formats.fits:setup
 export_python = glue.plugins.tools:setup
